+++ BUILD
@@ -27,10 +27,6 @@ load("//:deployment.bzl", deployment_github = "deployment")
 java_library(
     name = "common",
     srcs = glob([
-        "test/*.java",
-        "test/cluster/*.java",
-        "test/console/*.java",
-        "test/core/*.java",
         "collection/*.java",
         "concurrent/*.java",
         "concurrent/actor/*.java",
