+++ runner/BUILD
@@ -15,7 +15,12 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 
+load("@vaticle_bazel_distribution//github:rules.bzl", "deploy_github")
+load("@vaticle_bazel_distribution//maven:rules.bzl", "assemble_maven", "deploy_maven")
+load("@vaticle_dependencies//distribution/maven:version.bzl", "version")
+load("@vaticle_dependencies//library/maven:artifacts.bzl", "artifacts")
 load("@vaticle_dependencies//tool/checkstyle:rules.bzl", "checkstyle_test")
+load("@vaticle_dependencies//distribution:deployment.bzl", "deployment")
 
 java_library(
     name = "typedb-runner",
@@ -29,6 +34,25 @@ java_library(
         "@maven//:org_slf4j_slf4j_api"
     ],
     visibility = ["//visibility:public"],
+    tags = [
+        "maven_coordinates=com.vaticle.typedb:typedb-runner:{pom_version}",
+    ],
+)
+
+assemble_maven(
+    name = "assemble-maven",
+    target = ":typedb-runner",
+    workspace_refs = "@vaticle_typedb_common_workspace_refs//:refs.json",
+    version_overrides = version(artifacts_org = artifacts, artifacts_repo={}),
+    project_name = "TypeDB Runners",
+    project_description = "TypeDB Runner classes and tools",
+)
+
+deploy_maven(
+    name = "deploy-maven",
+    target = ":assemble-maven",
+    snapshot = deployment['maven.snapshot'],
+    release = deployment['maven.release']
 )
 
 filegroup(
